# abacusflow-portal/abacusflow-portal-web/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: abacusflow api
  version: 0.0.1
  description: 小算盘服务接口文档
servers:
  - url: http://localhost:8080
...
paths:
...
  /inventories:
    get:
      operationId: listBasicInventoriesPage
      tags:
        - inventory
      summary: Query inventories with pagination and filters
      parameters:
        - name: pageIndex
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 10
            maximum: 100
        - name: productCategoryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: productName
          in: query
          required: false
          schema:
            type: string
        - name: productType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProductType"
        - name: inventoryUnitCode
          in: query
          required: false
          schema:
            type: string
        - name: depotName
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Paged result of matching inventories
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageTemplate"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/BasicInventory"
  /inventories/{id}:
    get:
      operationId: getInventory
      tags:
        - inventory
      summary: Get inventory by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Inventory details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"
  /inventories/{id}/adjust-warning-line:
    post:
      operationId: adjustWarningLine
      tags:
        - inventory
      summary: Adjust safety and maximum stock levels
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - safetyStock
                - maxStock
              properties:
                safetyStock:
                  type: integer
                  format: int64
                  minimum: 0
                maxStock:
                  type: integer
                  format: int64
                  minimum: 0
      responses:
        "200":
          description: Warning line adjusted successfully
...
  /sale-orders:
    get:
      operationId: listBasicSaleOrdersPage
      tags:
        - transaction
      summary: List sale orders page
      parameters:
        - name: pageIndex
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 10
            maximum: 100
        - name: orderNo
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: customerName
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: productName
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of sale orders page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageTemplate"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/BasicSaleOrder"
    post:
      operationId: addSaleOrder
      tags:
        - transaction
      summary: Create a new sale order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleOrderInput"
      responses:
        "200":
          description: Sale order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleOrder"
  /sale-orders/{id}:
    get:
      operationId: getSaleOrder
      tags:
        - transaction
      summary: Get sale order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Sale order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleOrder"
  /sale-orders/{id}/complete:
    post:
      operationId: completeSaleOrder
      tags:
        - transaction
      summary: Complete a sale order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Sale order marked as completed
  /sale-orders/{id}/cancel:
    post:
      operationId: cancelSaleOrder
      tags:
        - transaction
      summary: Cancel a sale order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Sale order marked as canceled
  /sale-orders/{id}/reverse:
    post:
      operationId: reverseSaleOrder
      tags:
        - transaction
      summary: Reverse a sale order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Sale order marked as reversed
  /purchase-orders:
    get:
      operationId: listBasicPurchaseOrdersPage
      tags:
        - transaction
      summary: List purchase orders page
      parameters:
        - name: pageIndex
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 10
            maximum: 100
        - name: orderNo
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: supplierName
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: productName
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of purchase orders page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageTemplate"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/BasicPurchaseOrder"
    post:
      operationId: addPurchaseOrder
      tags:
        - transaction
      summary: Create a new purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderInput"
      responses:
        "200":
          description: Purchase order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags:
        - transaction
      summary: Get purchase order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Purchase order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}/complete:
    post:
      operationId: completePurchaseOrder
      tags:
        - transaction
      summary: Complete a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Purchase order marked as completeSaleOrder
  /purchase-orders/{id}/cancel:
    post:
      operationId: cancelPurchaseOrder
      tags:
        - transaction
      summary: Cancel a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Purchase order marked as canceled
  /purchase-orders/{id}/reverse:
    post:
      operationId: reversePurchaseOrder
      tags:
        - transaction
      summary: Reverse a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Purchase order marked as reversed
components:
  schemas:
...
    Inventory:
      type: object
      required:
        - id
        - productId
        - safetyStock
        - maxStock
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: 库存唯一标识
        productId:
          type: integer
          format: int64
          description: 产品ID
        safetyStock:
          type: integer
          format: int64
          description: 安全库存预警线
        maxStock:
          type: integer
          format: int64
          description: 最高库存预警线
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳
        updatedAt:
          type: integer
          format: int64
          description: 更新时间戳
    BasicInventory:
      type: object
      required:
        - id
        - productName
        - productType
        - initialQuantity
        - quantity
        - remainingQuantity
        - depotNames
        - safetyStock
        - maxStock
        - units
      properties:
        id:
          type: integer
          format: int64
          description: 库存ID
        productName:
          type: string
          description: 产品名称
        productType:
          $ref: "#/components/schemas/ProductType"
        productSpecification:
          type: string
          description: 产品规格
        productNote:
          type: string
          description: 产品备注
        initialQuantity:
          type: integer
          format: int64
          description: 初始总库存数量
        quantity:
          type: integer
          format: int64
          description: 现有总库存数量
        remainingQuantity:
          type: integer
          format: int64
          description: 可用总库存数量
        depotNames:
          description: 存储点列表
          type: array
          items:
            type: string
        safetyStock:
          type: integer
          format: int64
          description: 安全库存预警线
        maxStock:
          type: integer
          format: int64
          description: 最大库存预警线
        units:
          type: object
          items:
            $ref: "#/components/schemas/BasicInventoryUnit"

    # ============= InventoryUnit schemas =============
    BasicInventoryUnit:
      type: object
      required:
        - id
        - type
        - status
        - title
        - purchaseOrderNo
        - saleOrderNos
        - initialQuantity
        - quantity
        - remainingQuantity
        - unitPrice
        - receivedAt
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/InventoryUnitType"
        status:
          $ref: "#/components/schemas/InventoryUnitStatus"
        title:
          type: string
        purchaseOrderNo:
          type: string
          format: uuid
        saleOrderNos:
          type: array
          items:
            type: string
            format: uuid
        depotName:
          type: string
        initialQuantity:
          type: integer
          format: int64
          description: 初始库存数量
        quantity:
          type: integer
          format: int64
          description: 现有库存数量
        remainingQuantity:
          type: integer
          format: int64
          description: 可用库存数量
        unitPrice:
          type: number
          format: double
        receivedAt:
          type: integer
          format: int64
        batchCode:
          type: string
          format: uuid
          nullable: true
        serialNumber:
          type: string
          nullable: true
...
    InventoryUnitType:
      type: string
      description: 单元类型
      enum:
        - instance
        - batch
      x-enum-varnames:
        - instance
        - batch
      x-enum-descriptions:
        - 资产单元
        - 批次单元
    InventoryUnitStatus:
      type: string
      description: 单元类型
      enum:
        - normal
        - consumed
        - canceled
        - reversed
      x-enum-varnames:
        - normal
        - consumed
        - canceled
        - reversed
      x-enum-descriptions:
        - 初始状态
        - 已出库
        - 被人为取消
        - 因采购撤销等原因回退

    # ============= SaleOrder schemas =============
    SaleOrder:
      type: object
      required:
        - id
        - orderNo
        - customerId
        - status
        - totalAmount
        - orderItems
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: 销售订单唯一标识
        orderNo:
          type: string
          description: 订单编号
        customerId:
          type: integer
          format: int64
          description: 客户ID
        status:
          $ref: "#/components/schemas/OrderStatus"
        note:
          type: string
          nullable: true
          description: 备注
        orderDate:
          type: string
          format: date
          example: "1970-01-01"
          description: 订单日期
        orderItems:
          type: array
          items:
            $ref: "#/components/schemas/SaleOrderItem"
          description: 订单明细
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳
        updatedAt:
          type: integer
          format: int64
          description: 更新时间戳
    BasicSaleOrder:
      type: object
      required:
        - id
        - orderNo
        - customerName
        - status
        - totalAmount
        - totalQuantity
        - itemCount
        - orderDate
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: 销售订单ID
        orderNo:
          type: string
          description: 订单编号
        customerName:
          type: string
          description: 客户名称
        status:
          $ref: "#/components/schemas/OrderStatus"
        totalAmount:
          type: number
          format: double
          description: 订单总金额
        totalQuantity:
          type: integer
          format: int64
          description: 所有产品购买数量总和
        itemCount:
          type: integer
          description: 不同产品的种类数
        orderDate:
          type: string
          format: date
          example: "1970-01-01"
          description: 订单日期
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳
        autoCompleteDate:
          type: string
          format: date
          example: "1970-01-07"
          description: 自动完成天数
    CreateSaleOrderInput:
      type: object
      required:
        - customerId
        - orderItems
        - orderDate
      properties:
        customerId:
          type: integer
          format: int64
          description: 客户ID
        note:
          type: string
          nullable: true
          description: 备注
        orderDate:
          type: string
          format: date
          example: "1970-01-01"
          description: 订单日期
        orderItems:
          type: array
          items:
            $ref: "#/components/schemas/SaleOrderItemInput"
          minItems: 1
          description: 订单明细

    # ============= PurchaseOrder schemas =============
    PurchaseOrder:
      type: object
      required:
        - id
        - orderNo
        - supplierId
        - status
        - totalAmount
        - orderItems
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: 采购订单唯一标识
        orderNo:
          type: string
          description: 订单编号
        supplierId:
          type: integer
          format: int64
          description: 供应商ID
        status:
          $ref: "#/components/schemas/OrderStatus"
        note:
          type: string
          nullable: true
          description: 备注
        orderDate:
          type: string
          format: date
          example: "1970-01-01"
          description: 订单日期
        orderItems:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
          description: 订单明细
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳
        updatedAt:
          type: integer
          format: int64
          description: 更新时间戳
    BasicPurchaseOrder:
      type: object
      required:
        - id
        - orderNo
        - supplierName
        - status
        - totalAmount
        - totalQuantity
        - itemCount
        - orderDate
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: 采购订单ID
        orderNo:
          type: string
          description: 订单编号
        supplierName:
          type: string
          description: 供应商名称
        status:
          $ref: "#/components/schemas/OrderStatus"
        totalAmount:
          type: number
          format: double
          description: 订单总金额
        totalQuantity:
          type: integer
          format: int64
          description: 所有产品购买数量总和
        itemCount:
          type: integer
          description: 不同产品的种类数
        orderDate:
          type: string
          format: date
          example: "1970-01-01"
          description: 订单日期
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳
        autoCompleteDate:
          type: string
          format: date
          example: "1970-01-07"
          description: 自动完成天数
    CreatePurchaseOrderInput:
      type: object
      required:
        - supplierId
        - orderDate
        - orderItems
      properties:
        supplierId:
          type: integer
          format: int64
          description: 供应商ID
        note:
          type: string
          nullable: true
          description: 备注
        orderDate:
          type: string
          format: date
          example: "1970-01-01"
          description: 订单日期
        orderItems:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItemInput"
          minItems: 1
          description: 订单明细

    # ============= OrderItem schemas =============
    SaleOrderItem:
      type: object
      required:
        - id
        - inventoryUnitId
        - quantity
        - unitPrice
        - discountedPrice
        - subtotal
      properties:
        id:
          type: integer
          format: int64
          description: 订单明细ID
        inventoryUnitId:
          type: integer
          format: int64
          description: 库存单元ID
        quantity:
          type: integer
          description: 数量
        unitPrice:
          type: number
          format: double
          description: 单价
        discountedPrice:
          type: number
          format: double
          description: 折扣价
        subtotal:
          type: number
          format: double
          description: 小计
    SaleOrderItemInput:
      type: object
      required:
        - inventoryUnitId
        - quantity
        - unitPrice
      properties:
        inventoryUnitId:
          type: integer
          format: int64
          description: 库存单元ID
        quantity:
          type: integer
          minimum: 1
          description: 数量
        unitPrice:
          type: number
          format: double
          minimum: 0
          description: 单价
        discountFactor:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: 折扣率
    PurchaseOrderItem:
      type: object
      required:
        - id
        - productId
        - quantity
        - unitPrice
        - subtotal
      properties:
        id:
          type: integer
          format: int64
          description: 订单明细ID
        productId:
          type: integer
          format: int64
          description: 产品ID
        quantity:
          type: integer
          description: 数量
        unitPrice:
          type: number
          format: double
          description: 单价
        serialNumber:
          type: string
        subtotal:
          type: number
          format: double
          description: 小计
    PurchaseOrderItemInput:
      type: object
      required:
        - productId
        - quantity
        - unitPrice
      properties:
        productId:
          type: integer
          format: int64
          description: 产品ID
        quantity:
          type: integer
          minimum: 1
          description: 数量
        unitPrice:
          type: number
          format: double
          minimum: 0
          description: 单价
        serialNumber:
          type: string
...
    ProductType:
      type: string
      description: 产品类型
      enum:
        - material
        - asset
      x-enum-descriptions:
        - 普通产品
        - 资产
    OrderStatus:
      type: string
      description: 订单状态
      enum:
        - pending
        - completed
        - canceled
        - reversed
      x-enum-varnames:
        - pending
        - completed
        - canceled
        - reversed
      x-enum-descriptions:
        - 待处理
        - 已完成
        - 已取消
        - 已撤回
    PageTemplate:
      type: object
      required:
        - content
        - totalElements
        - number
        - size
      properties:
        content:
          type: array
          items:
            type: string
        totalElements:
          type: integer
          format: int64
        number:
          type: integer
          description: 当前页码，从 0 开始
        size:
          type: integer
          description: 每页条数
...